refactor(vault): :recycle: move sync CLI pure helpers to lib/sync/cli-helpers.mjs
Drops the import.meta.main guard (which would silently no-op the bin on
older Node); helpers now live in a lib module the test imports directly,
so bin/cli.mjs dispatches unconditionally as before.
