+++ plugins/agent-id-vault/bin/cli.mjs
@@ -89,6 +89,7 @@ import { announceBeacon, listenForBeacons } from "../lib/sync/discovery.mjs";
 import { withVaultLock } from "../lib/sync/lock.mjs";
 import { runSyncSession } from "../lib/sync/protocol.mjs";
 import { ensureSyncMeta, loadSyncIdentity, pinDevice, revokeDevice } from "../lib/sync/trust.mjs";
+import { redactSecretFields, parseListenPort } from "../lib/sync/cli-helpers.mjs";
 
 // ─── Input helpers ──────────────────────────────────────────────────────────────
 
@@ -1227,36 +1228,6 @@ async function cmdExec() {
 
 // ─── sync ───────────────────────────────────────────────────────────────────────
 
-// Redact every secret-bearing field on a credential record before it is printed.
-// Used by `sync resolve` (no --restore) so the journaled *losing* record can be
-// inspected without leaking refreshToken/clientSecret/accessToken (oauth2),
-// password/totpSecret (login), privateKey (evm), secretSeed (solana), etc.
-// Pure: returns a shallow copy; the input is left untouched.
-export function redactSecretFields(rec) {
-  const redacted = { ...rec };
-  for (const f of SECRET_FIELDS) {
-    if (redacted[f] != null) redacted[f] = "(redacted — use --restore)";
-  }
-  return redacted;
-}
-
-// Validate the --listen --port flag. parseFlags yields boolean `true` for a bare
-// flag (last, or followed by another --flag); Number(true) === 1 would silently
-// bind port 1, so a boolean or non-digit value must be rejected. Returns 0 when
-// absent (ephemeral OS-assigned port). Throws Error (message names --port) so
-// the caller can outputError.
-export function parseListenPort(flags) {
-  if (flags.port == null) return 0;
-  if (typeof flags.port !== "string" || !/^[0-9]+$/.test(flags.port)) {
-    throw new Error("--port must be a number (0–65535)");
-  }
-  const port = Number(flags.port);
-  if (!Number.isInteger(port) || port < 0 || port > 65535) {
-    throw new Error("--port must be a number (0–65535)");
-  }
-  return port;
-}
-
 function promptApproval(peer) {
   if (process.env.AGENT_ID_SYNC_AUTOAPPROVE === "1") {
     // Escape hatch for automated e2e (Docker) — never silent: surface a loud
@@ -1588,8 +1559,4 @@ for (const k of Object.keys(commands)) {
   };
 }
 
-// Run the CLI only when invoked directly (`node cli.mjs …`), not when imported
-// (e.g. by the unit tests that exercise the pure helpers below).
-if (import.meta.main) {
-  runCli({ commands, printHelp });
-}
+runCli({ commands, printHelp });
